SMART BLIND STICK USING IOT
The ultrasonic sensor is used to detect the presence of obstacle and calculates the distance between the
source and destination.Through this smart blind stick, visually impaired people will have so much of assistance. In this
technology- controlled world, where people strive to live independently, this project proposes an ultrasonic
stick for blindpeople to help them gain personal independence. Since this is economical and not bulky, one
can make use ofit easily.