SMART BLIND STICK USING IOT
The main objective of our project is to provide a sound-based assistance to blind people. Here we are trying to develop a system that helps blind person to travel independently and works efficiently. Through this smart blind stick, visually impaired people will have so much of assistance. In this technology- controlled world, where people strive to live independently, this project proposes an ultrasonic stick for blind people to help them gain personal independence. Since this is economical and not bulky, one
can make use of it easily.